@@ -4,7 +4,7 @@ import 'package:qr_scanner_app/service/controllers/common_controllers/bottom_bar
 import 'package:qr_scanner_app/utils/colors.dart';
 
 class HomePage extends StatelessWidget {
-  const HomePage({Key? key}) : super(key: key);
+  const HomePage({super.key});
 
   @override
   Widget build(BuildContext context) {
